backend/requirements.txt
fastapi
uvicorn
python-docx
bcrypt
dashscope
python-dotenv
python-multipart